+++ models/segnn/segnn.py
@@ -10,53 +10,104 @@ from .instance_norm import InstanceNorm
 
 
 class SEGNN(nn.Module):
-    """ Steerable E(3) equivariant message passing network """
+    """Steerable E(3) equivariant message passing network"""
 
-    def __init__(self, input_irreps, hidden_irreps, output_irreps, edge_attr_irreps, node_attr_irreps, num_layers, norm=None, pool='avg', task='graph', init="kaiming_uniform", additional_message_irreps=None):
+    def __init__(
+        self,
+        input_irreps,
+        hidden_irreps,
+        output_irreps,
+        edge_attr_irreps,
+        node_attr_irreps,
+        num_layers,
+        norm=None,
+        pool="avg",
+        task="graph",
+        additional_message_irreps=None,
+    ):
         super().__init__()
         self.task = task
         # Create network, embedding first
-        self.embedding_layer_1 = O3TensorProductSwishGate(input_irreps, hidden_irreps, node_attr_irreps, init=init)
-        self.embedding_layer_2 = O3TensorProduct(hidden_irreps, hidden_irreps, node_attr_irreps, init=init)
+        # self.embedding_layer_1 = O3TensorProductSwishGate(
+        #     input_irreps, hidden_irreps, node_attr_irreps
+        # )
+        # self.embedding_layer_2 = O3TensorProduct(
+        #     hidden_irreps, hidden_irreps, node_attr_irreps
+        # )
+
+        self.embedding_layer = O3TensorProduct(
+            input_irreps, hidden_irreps, node_attr_irreps
+        )
 
         # Message passing layers.
         layers = []
         for i in range(num_layers):
-            layers.append(SEGNNLayer(hidden_irreps, hidden_irreps, hidden_irreps, edge_attr_irreps,
-                                     node_attr_irreps, norm=norm, init=init, additional_message_irreps=additional_message_irreps))
+            layers.append(
+                SEGNNLayer(
+                    hidden_irreps,
+                    hidden_irreps,
+                    hidden_irreps,
+                    edge_attr_irreps,
+                    node_attr_irreps,
+                    norm=norm,
+                    additional_message_irreps=additional_message_irreps,
+                )
+            )
         self.layers = nn.ModuleList(layers)
 
         # Prepare for output irreps, since the attrs will disappear after pooling
-        if task == 'graph':
-            pooled_irreps = (output_irreps*hidden_irreps.num_irreps).simplify().sort().irreps
-            self.pre_pool1 = O3TensorProductSwishGate(hidden_irreps, hidden_irreps, node_attr_irreps, init=init)
-            self.pre_pool2 = O3TensorProduct(hidden_irreps, pooled_irreps, node_attr_irreps, init=init)
-            self.post_pool1 = O3TensorProductSwishGate(pooled_irreps, pooled_irreps, init=init)
-            self.post_pool2 = O3TensorProduct(pooled_irreps, output_irreps, init=init)
+        if task == "graph":
+            pooled_irreps = (
+                (output_irreps * hidden_irreps.num_irreps).simplify().sort().irreps
+            )
+            self.pre_pool1 = O3TensorProductSwishGate(
+                hidden_irreps, hidden_irreps, node_attr_irreps
+            )
+            self.pre_pool2 = O3TensorProduct(
+                hidden_irreps, pooled_irreps, node_attr_irreps
+            )
+            self.post_pool1 = O3TensorProductSwishGate(pooled_irreps, pooled_irreps)
+            self.post_pool2 = O3TensorProduct(pooled_irreps, output_irreps)
             self.init_pooler(pool)
-        elif task == 'node':
-            self.pre_pool1 = O3TensorProductSwishGate(hidden_irreps, hidden_irreps, node_attr_irreps, init=init)
-            self.pre_pool2 = O3TensorProduct(hidden_irreps, output_irreps, node_attr_irreps, init=init)
+        elif task == "node":
+            self.pre_pool1 = O3TensorProductSwishGate(
+                hidden_irreps, hidden_irreps, node_attr_irreps
+            )
+            self.pre_pool2 = O3TensorProduct(
+                hidden_irreps, output_irreps, node_attr_irreps
+            )
 
     def init_pooler(self, pool):
-        """ Initialise pooling mechanism """
+        """Initialise pooling mechanism"""
         if pool == "avg":
             self.pooler = global_mean_pool
         elif pool == "sum":
             self.pooler = global_add_pool
 
     def catch_isolated_nodes(self, graph):
-        """ Isolated nodes should also obtain attributes """
-        if (graph.contains_isolated_nodes() and graph.edge_index.max().item() + 1 != graph.num_nodes):
+        """Isolated nodes should also obtain attributes"""
+        if (
+            graph.contains_isolated_nodes()
+            and graph.edge_index.max().item() + 1 != graph.num_nodes
+        ):
             nr_add_attr = graph.num_nodes - (graph.edge_index.max().item() + 1)
-            add_attr = graph.node_attr.new_tensor(np.zeros((nr_add_attr, node_attr.shape[-1])))
+            add_attr = graph.node_attr.new_tensor(
+                np.zeros((nr_add_attr, node_attr.shape[-1]))
+            )
             graph.node_attr = torch.cat((graph.node_attr, add_attr), -2)
         # Trivial irrep value should always be 1 (is automatically so for connected nodes, but isolated nodes are now 0)
-        graph.node_attr[:, 0] = 1.
+        graph.node_attr[:, 0] = 1.0
 
     def forward(self, graph):
-        """ SEGNN forward pass """
-        x, pos, edge_index, edge_attr, node_attr, batch = graph.x, graph.pos, graph.edge_index, graph.edge_attr, graph.node_attr, graph.batch
+        """SEGNN forward pass"""
+        x, pos, edge_index, edge_attr, node_attr, batch = (
+            graph.x,
+            graph.pos,
+            graph.edge_index,
+            graph.edge_attr,
+            graph.node_attr,
+            graph.batch,
+        )
         try:
             additional_message_features = graph.additional_message_features
         except AttributeError:
@@ -65,18 +116,21 @@ class SEGNN(nn.Module):
         self.catch_isolated_nodes(graph)
 
         # Embed
-        x = self.embedding_layer_1(x, node_attr)
-        x = self.embedding_layer_2(x, node_attr)
+        # x = self.embedding_layer_1(x, node_attr)
+        # x = self.embedding_layer_2(x, node_attr)
+        x = self.embedding_layer(x, node_attr)
 
         # Pass messages
         for layer in self.layers:
-            x = layer(x, edge_index, edge_attr, node_attr, batch, additional_message_features)
+            x = layer(
+                x, edge_index, edge_attr, node_attr, batch, additional_message_features
+            )
 
         # Pre pool
         x = self.pre_pool1(x, node_attr)
         x = self.pre_pool2(x, node_attr)
 
-        if self.task == 'graph':
+        if self.task == "graph":
             # Pool over nodes
             x = self.pooler(x, batch)
 
@@ -87,25 +141,41 @@ class SEGNN(nn.Module):
 
 
 class SEGNNLayer(MessagePassing):
-    """ E(3) equivariant message passing layer. """
+    """E(3) equivariant message passing layer."""
 
-    def __init__(self, input_irreps, hidden_irreps, output_irreps, edge_attr_irreps, node_attr_irreps, norm=None, init="kaiming_uniform", additional_message_irreps=None):
+    def __init__(
+        self,
+        input_irreps,
+        hidden_irreps,
+        output_irreps,
+        edge_attr_irreps,
+        node_attr_irreps,
+        norm=None,
+        additional_message_irreps=None,
+    ):
         super().__init__(node_dim=-2, aggr="add")
         self.hidden_irreps = hidden_irreps
 
-        message_input_irreps = (2*input_irreps + additional_message_irreps).simplify()
+        message_input_irreps = (2 * input_irreps + additional_message_irreps).simplify()
         update_input_irreps = (input_irreps + hidden_irreps).simplify()
 
         self.message_layer_1 = O3TensorProductSwishGate(
-            message_input_irreps, hidden_irreps, edge_attr_irreps, init=init)
-        self.message_layer_2 = O3TensorProductSwishGate(hidden_irreps, hidden_irreps, edge_attr_irreps, init=init)
-        self.update_layer_1 = O3TensorProductSwishGate(update_input_irreps, hidden_irreps, node_attr_irreps, init=init)
-        self.update_layer_2 = O3TensorProduct(hidden_irreps, hidden_irreps, node_attr_irreps, init=init)
+            message_input_irreps, hidden_irreps, edge_attr_irreps
+        )
+        self.message_layer_2 = O3TensorProductSwishGate(
+            hidden_irreps, hidden_irreps, edge_attr_irreps
+        )
+        self.update_layer_1 = O3TensorProductSwishGate(
+            update_input_irreps, hidden_irreps, node_attr_irreps
+        )
+        self.update_layer_2 = O3TensorProduct(
+            hidden_irreps, hidden_irreps, node_attr_irreps
+        )
 
         self.setup_normalisation(norm)
 
     def setup_normalisation(self, norm):
-        """ Set up normalisation, either batch or instance norm """
+        """Set up normalisation, either batch or instance norm"""
         self.norm = norm
         self.feature_norm = None
         self.message_norm = None
@@ -116,10 +186,23 @@ class SEGNNLayer(MessagePassing):
         elif norm == "instance":
             self.feature_norm = InstanceNorm(self.hidden_irreps)
 
-    def forward(self, x, edge_index, edge_attr, node_attr, batch, additional_message_features=None):
-        """ Propagate messages along edges """
-        x = self.propagate(edge_index, x=x, node_attr=node_attr, edge_attr=edge_attr,
-                           additional_message_features=additional_message_features)
+    def forward(
+        self,
+        x,
+        edge_index,
+        edge_attr,
+        node_attr,
+        batch,
+        additional_message_features=None,
+    ):
+        """Propagate messages along edges"""
+        x = self.propagate(
+            edge_index,
+            x=x,
+            node_attr=node_attr,
+            edge_attr=edge_attr,
+            additional_message_features=additional_message_features,
+        )
         # Normalise features
         if self.feature_norm:
             if self.norm == "batch":
@@ -129,7 +212,7 @@ class SEGNNLayer(MessagePassing):
         return x
 
     def message(self, x_i, x_j, edge_attr, additional_message_features):
-        """ Create messages """
+        """Create messages"""
         if additional_message_features is None:
             input = torch.cat((x_i, x_j), dim=-1)
         else:
@@ -143,7 +226,7 @@ class SEGNNLayer(MessagePassing):
         return message
 
     def update(self, message, x, node_attr):
-        """ Update note features """
+        """Update note features"""
         input = torch.cat((x, message), dim=-1)
         update = self.update_layer_1(input, node_attr)
         update = self.update_layer_2(update, node_attr)
